# weyyak_gcp/staging/api/content/docs/swagger.yaml
definitions:
...
  director.Director:
    properties:
      arabicName:
        type: string
      englishName:
        type: string
    type: object
...
  singer.Singer:
    properties:
      arabicName:
        type: string
      englishName:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /api/directors:
    get:
      consumes:
      - application/json
      description: get list of all Director's
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/director.Director'
            type: array
      security:
      - Authorization: []
      summary: Show a list of all Director's
      tags:
      - Directors
    put:
      consumes:
      - application/json
      description: Create New Director
      parameters:
      - description: Raw JSON string
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/director.Director'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/director.Director'
            type: array
      summary: Create New Director
      tags:
      - Directors
...
  /api/singers:
    get:
      consumes:
      - application/json
      description: get list of all Singer's
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/singer.Singer'
            type: array
      security:
      - Authorization: []
      summary: Show a list of all Singer's
      tags:
      - Singers
    put:
      consumes:
      - application/json
      description: Create New Singer
      parameters:
      - description: Raw JSON string
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/singer.Singer'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/singer.Singer'
            type: array
      summary: Create New Singer
      tags:
      - Singers
...
swagger: "2.0"